move type selection dialog in separate function add ctrl+n key to create new graph make button to undo addition
