+++ app/src/main/kotlin/viewmodel/MainScreenViewModel.kt
@@ -12,7 +12,11 @@ import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
 import model.Edge
 import algo.keyvertex.KeyVertexFinder
+import model.graphs.DirWeightGraph
+import model.graphs.DirectedGraph
 import model.graphs.EmptyGraph
+import model.graphs.UndirWeightGraph
+import model.graphs.UndirectedGraph
 import viewmodel.ColorList
 import java.io.File
 import java.util.Vector
@@ -72,6 +76,8 @@ class MainScreenViewModel<K, V>(graphViewModel: GraphViewModel<K, V>) {
     val showKeyVertexDialog = mutableStateOf(false)
     val showKeyVerticesResult = mutableStateOf(false)
     val keyVerticesCount = mutableStateOf(0)
+
+
 
     val planarAlgos: (Planar) -> Unit = {
         clean()
@@ -459,5 +465,16 @@ class MainScreenViewModel<K, V>(graphViewModel: GraphViewModel<K, V>) {
             errorText.value=e.message.toString()
             error.value=true
         }
+    }
+
+    fun graphTypeSelection(selectedOption: String, graphs: List<String>) {
+        val graph = when (selectedOption) {
+            graphs[0] -> UndirectedGraph<K, V>()
+            graphs[1] -> UndirWeightGraph()
+            graphs[2] -> DirWeightGraph()
+            else -> DirectedGraph()
+        }
+        viewModel = GraphViewModel(graph)
+        graphType.value=false
     }
 }
